fix: create a notification hook for adding noti
