+++ src/pages/TestNotification.tsx
@@ -1,23 +1,8 @@
 import { NotificationStack } from '@/components/Notification'
-import { NotificationType } from '@/libs/constants'
-import { useNotificationStore } from '@/stores/notification'
+import { useAddNotification } from '@/hooks/useNotification'
 
 const TestNotification = () => {
-  const [notifications, setNotifications] = useNotificationStore((state) => [
-    state.notifications,
-    state.setNotifications,
-  ])
-
-  const handleAddNotification = (type: NotificationType, message: string) => {
-    const newNotifications = new Map(notifications)
-    const id = Math.random()
-
-    newNotifications.set(id, { id, type, message })
-    if (newNotifications.size > 3) {
-      newNotifications.delete(newNotifications.keys().next().value)
-    }
-    setNotifications(newNotifications)
-  }
+  const handleAddNotification = useAddNotification()
 
   return (
     <div style={{ display: 'flex', flexDirection: 'column', margin: '400px' }}>
